+++ crates/miden-lib/src/transaction/memory.rs
@@ -16,7 +16,7 @@ pub type StorageSlot = u8;
 //
 // | Section            | Start address, pointer (word pointer) | End address, pointer (word pointer) | Comment                                    |
 // | ------------------ | :-----------------------------------: | :---------------------------------: | ------------------------------------------ |
-// | Bookkeeping        | 0 (0)                                 | 287 (71)                            |                                            |
+// | Bookkeeping        | 0 (0)                                 | 96 (24)                             |                                            |
 // | Global inputs      | 400 (100)                             | 439 (109)                           |                                            |
 // | Block header       | 800 (200)                             | 843 (210)                           |                                            |
 // | Partial blockchain | 1_200 (300)                           | 1_331? (332?)                       |                                            |
@@ -91,13 +91,20 @@ pub const OUTPUT_VAULT_ROOT_PTR: MemoryAddress = 16;
 /// The memory address at which the native account's new code commitment is stored.
 pub const NEW_CODE_ROOT_PTR: MemoryAddress = 20;
 
+/// The memory address at which the dirty flag of the storage commitment of the native account is
+/// stored.
+///
+/// This binary flag specifies whether the commitment is outdated: it holds 1 if some changes were
+/// made to the account storage since the last re-computation, and 0 otherwise.
+pub const NATIVE_ACCT_STORAGE_COMMITMENT_DIRTY_FLAG_PTR: MemoryAddress = 24;
+
 /// The memory address at which the transaction expiration block number is stored.
-pub const TX_EXPIRATION_BLOCK_NUM_PTR: MemoryAddress = 24;
+pub const TX_EXPIRATION_BLOCK_NUM_PTR: MemoryAddress = 28;
 
 /// The memory address at which the pointer to the stack element containing the pointer to the
 /// currently active account data is stored.
 ///
-/// The stack starts at the address `29`. Stack has a length of `64` elements meaning that the
+/// The stack starts at the address `33`. Stack has a length of `64` elements meaning that the
 /// maximum depth of FPI calls is `63` — the first slot is always occupied by the native account
 /// data pointer.
 ///
@@ -105,9 +112,9 @@ pub const TX_EXPIRATION_BLOCK_NUM_PTR: MemoryAddress = 24;
 /// ┌───────────────┬────────────────┬───────────────────┬─────┬────────────────────┐
 /// │ STACK TOP PTR │ NATIVE ACCOUNT │ FOREIGN ACCOUNT 1 │ ... │ FOREIGN ACCOUNT 63 │
 /// ├───────────────┼────────────────┼───────────────────┼─────┼────────────────────┤
-///        28               29                30                         92
+///        32               33                34                         96
 /// ```
-pub const ACCOUNT_STACK_TOP_PTR: MemoryAddress = 28;
+pub const ACCOUNT_STACK_TOP_PTR: MemoryAddress = 32;
 
 // GLOBAL INPUTS
 // ------------------------------------------------------------------------------------------------